==> gene_expression/2_Acer_PCA.Rmd <==
---
title: "Acer_PCA"
author: "allyson_demerlis"
date: "2023-11-24"
output: html_document
---

```{r}
library(DESeq2)
library(tidyverse)
library(factoextra)
library(cowplot)
library(limma)
library(ggforce)

load("RData_files/initial_fullddsdesigncountsVsdcounts.RData")
```

# all data

```{r}
pca <- prcomp(t(assay(Vsd)))
summary(pca)
fviz_eig(pca)
```

Note: the data going into these PCAs is after the outlier removal step using arrayQualityMetrics
```{r}
plotPCA(Vsd, intgroup = c("Location"))
plotPCA(Vsd, intgroup = c("Genotype"))
plotPCA(Vsd, intgroup = c("Location", "Genotype"))

vst_PCAdata <- plotPCA(Vsd, intgroup = c("Location", "Genotype"), returnData = TRUE)
percentVar <- round(100*attr(vst_PCAdata, "percentVar")) 

ggplot(vst_PCAdata, aes(PC1, PC2, color=Location, shape=Genotype)) +
   geom_point(size=3) +
   xlab(paste0("PC1: ",percentVar[1],"% variance")) +
   ylab(paste0("PC2: ",percentVar[2],"% variance")) +
   ggtitle("A.cervicornis") +
   theme_classic() + #Set background color
   theme(panel.border = element_blank(), # Set border
         axis.line = element_line(colour = "black"), #Set axes color
         plot.background=element_blank())   +
  scale_color_manual(values = c("CCC" = "orange", "nursery" = "darkblue"))
#ggsave("PCA_allgenets.pdf")
```
