==> gene_expression/6_WGCNA_GO.Rmd <==
---
title: "WGCNA_GO"
author: "Allyson"
date: "`r Sys.Date()`"
output: html_document
---

Following https://rpubs.com/LeonYu/995707 and [Dr Chrstine Roper's GitHub](https://github.com/ChristineRoper/Mangrove-coral-thermotolerance-retained/blob/main/RNA%20Seq/4-Functional_annotation/Functional_annotation.R).

# install libraries
```{r}
# BiocManager::install("Rgraphviz")
# BiocManager::install("topGO")
# BiocManager::install("clusterProfiler")
# BiocManager::install("GOplot")
# BiocManager::install("GSEABase")
# BiocManager::install("GO.db")
# BiocManager::install("rrvgo")
```

# load libraries and R Data files
```{r}
library(tidyverse)
library(dplyr)
library(ggplot2)
library(reshape2)
library(topGO)
library(plyr)
library(tidyr)
library(scales)
library(stringr)
library(GSEABase)
library(rrvgo)

load("Rdata_files/initial_fullddsdesigncountsVsdcounts.RData")
load("Rdata_files/realModels_Acer.RData")
load("Rdata_files/vsd.RData")
load("Rdata_files/pvals.RData")

geneID2GO <- read.delim("~/OneDrive - University of Miami/PhD/NOAA AOML Coral Program/chapter2_UrbanCorals/Acropora_iso2go.tab", header = FALSE, stringsAsFactors = FALSE)
colnames(geneID2GO) <- c("gene", "GO")
length(geneID2GO) #7995

# Split on semicolon
geneID2GO <- setNames(strsplit(geneID2GO$GO, ";"), geneID2GO$gene)

# Background = all genes from the mapping
allgenes <- names(geneID2GO)
```

# Make a list of genes and modules
```{r}
# Path to your folder containing the CSV files
module_gene_lists_subfolder <- "4_WGCNA/module_gene_lists"

# List all files ending with "_genelist.csv" in the folder
file_list <- list.files(path = module_gene_lists_subfolder, 
                        pattern = "_genelist\\.csv$", 
                        full.names = TRUE)

# Read all files and add a 'module' column with the file name
all_genes <- file_list %>%
  set_names() %>%                    # name the list with full paths
    map_dfr(~ read_csv(.x) %>%
            mutate(module = str_remove(tools::file_path_sans_ext(basename(.x)),
                                       "_genelist$")),
          .id = NULL)

# Check result
head(all_genes)

all_genes %>% 
  dplyr::select(gene, module) -> all_genes
```


Loop to run all modules
```{r}
# Background universe: all unique genes across all modules
allgenes <- unique(all_genes$gene)

# List of modules to iterate over
modules <- unique(all_genes$module)

# Initialize results container
GO_results <- list()

# Loop through each module
for(mod in modules) {
  
  # Genes in this specific module
  module_genes <- all_genes$gene[all_genes$module == mod]
  
  # Build geneList for topGO (0/1 factor)
  geneList <- factor(as.integer(allgenes %in% module_genes),
                     levels = c(0,1))
  names(geneList) <- allgenes
  
  #Build the topGOdata matrix for each ontology categories
  GOdata_BP <- new("topGOdata", ontology = "BP", allGenes = geneList,
                    annot = annFUN.gene2GO, gene2GO = geneID2GO,nodeSize=5)

  GOdata_CC <- new("topGOdata", ontology = "CC", allGenes = geneList,
                    annot = annFUN.gene2GO, gene2GO = geneID2GO,nodeSize=5)

  GOdata_MF <- new("topGOdata", ontology = "MF", allGenes = geneList,
                    annot = annFUN.gene2GO, gene2GO = geneID2GO, nodeSize=5)
  
  #Perform enrichment analysis 
  
  # run weighted algorithm as classic doesnt take into consideration GO hierarchy so could overrepresent enrichment
  resultTopGO.BP <- runTest(GOdata_BP, algorithm = "weight01", statistic = "fisher")
  resultTopGO.CC <- runTest(GOdata_CC, algorithm = "weight01", statistic = "fisher" )
  resultTopGO.MF <- runTest(GOdata_MF, algorithm = "weight01", statistic = "fisher" )
  
  allBPGO = genesInTerm(GOdata_BP)
  allCCGO = genesInTerm(GOdata_CC)
  allMFGO = genesInTerm(GOdata_MF)
  
  ### Load function for extracting genes associated with a GO ID
  BP_ANNOTATION = lapply(allBPGO,function(x) x[x %in% module_genes])
  CC_ANNOTATION = lapply(allCCGO,function(x) x[x %in% module_genes])
  MF_ANNOTATION = lapply(allMFGO,function(x) x[x %in% module_genes])
  
  #Transform data and save into csv
  # Create results tables (numChar parameter controls term description length)
BP_Res <- GenTable(GOdata_BP, weightFisher = resultTopGO.BP, topNodes = length(allBPGO), numChar = 1000)
BP_Res$Class <- "Biological_Process"
BP_Res$GeneList <- vapply(BP_Res$GO.ID, function(go) {
  str_c(BP_ANNOTATION[[go]], collapse = ",")
}, character(1))

CC_Res <- GenTable(GOdata_CC, weightFisher = resultTopGO.CC, topNodes = length(allCCGO), numChar = 1000)
CC_Res$Class <- "Cellular_Component"
CC_Res$GeneList <- vapply(CC_Res$GO.ID, function(go) {
  str_c(CC_ANNOTATION[[go]], collapse = ",")
}, character(1))

MF_Res <- GenTable(GOdata_MF, weightFisher = resultTopGO.MF, topNodes = length(allMFGO), numChar = 1000)
MF_Res$Class <- "Molecular_Function"
MF_Res$GeneList <- vapply(MF_Res$GO.ID, function(go) {
  str_c(MF_ANNOTATION[[go]], collapse = ",")
}, character(1))

  GO_combine <- bind_rows(MF_Res, CC_Res, BP_Res)
  GO_combine$weightFisher <- as.numeric(GO_combine$weightFisher)
  GO_combine <- GO_combine[GO_combine$weightFisher < 0.01, ]
  GO_combine$module <- mod
  
  GO_results[[mod]] <- GO_combine
}

str(GO_results)

GO_results$darkmagenta

GO_results$mediumpurple3

write_csv(GO_results, "4_WGCNA/GO_enrichment_all_modules_allterms.csv")
```


GOslim
```{r}
#change Class types to ontology so it matches GOslim
GO_results %>% 
  mutate(Ontology = case_when(
    Class == "Biological_Process" ~ "BP",
    Class == "Molecular_Function" ~ "MF",
    Class == "Cellular_Component" ~ "CC"
  )) -> GO_results

# Load the GO Slim file
slim <- getOBOCollection("goslim_generic.obo") #download here (https://release.geneontology.org/2024-09-08/ontology/subsets/index.html#:~:text=281%20KB-,goslim_generic.obo,-117%20KB)

slim_ids <- GSEABase::ids(slim) # gets slim IDs

# Helper function: map slim → detailed terms → genes
goWide <- function(slim, genes, GO_ontology) {
    GOTERM <- getAnnMap("TERM", "GO")
    terms <- mget(slim_ids, GOTERM, ifnotfound = NA)
    terms <- terms[vapply(terms, Ontology, character(1)) == GO_ontology]
    slim_ids_o <- names(terms)

    OFFSPRING <- switch(GO_ontology,
        MF = getAnnMap("MFOFFSPRING", "GO"),
        BP = getAnnMap("BPOFFSPRING", "GO"),
        CC = getAnnMap("CCOFFSPRING", "GO"),
        stop("GO_ontology must be 'MF', 'BP', or 'CC'")
    )

    slim_to_gene <- mget(slim_ids_o, OFFSPRING, ifnotfound = NA)
    slim_to_gene <- slim_to_gene[!is.na(slim_to_gene)]
    gene_to_slim <- revmap(slim_to_gene)[genes]
    gene_to_slim <- gene_to_slim[!vapply(gene_to_slim, is.null, logical(1))]

    if (length(gene_to_slim) == 0) {
        return(c())
    }

    return(names(gene_to_slim)[sapply(gene_to_slim, function(x) any(grepl(slim, unlist(x))))])
}

all_slim_results <- list()

# Loop through modules in GO_results
for (mod in unique(GO_results$module)) {
    message("Processing module: ", mod)

    # Subset one module’s enriched GO terms
    module_df <- GO_results %>% filter(module == mod)
    GO_IDs <- module_df$GO.ID
    GO_child_term <- module_df$Term

    # Wrap in GOCollection
    collection <- GOCollection(GO_IDs)

    # Run GOslim for each ontology
    for (ontology in c("BP", "MF", "CC")) {
          tmp <- goSlim(GOCollection(module_df$GO.ID), slim, ontology) %>%
            rownames_to_column("GO") %>%
            mutate(
                module = mod,
                Ontology = ontology
            ) %>%
            rowwise() %>%
                mutate(GOTerms = paste(goWide(GO, GO_IDs, ontology), collapse = ", ")) %>% # adding new column for the GO terms
            filter(Count > 0)

        all_slim_results[[paste(mod, ontology, sep = "_")]] <- tmp

    }
}

# Combine everything into one tibble
all_slim_results <- bind_rows(all_slim_results)
# save to csv
write_csv(all_slim_results, "4_WGCNA/GOslim_enrichment_all_modules.csv")

```


```{r}
GO_results <- read_csv("4_WGCNA/GO_enrichment_all_modules_allterms.csv")
GO_slim <- read_csv("4_WGCNA/GOslim_enrichment_all_modules.csv")

GO_slim %>% 
  dplyr::rename(ParentTerm = GO) %>% 
  separate_rows(GOTerms, sep = ",\\s*") %>%
  mutate(GOTerms = ifelse(is.na(GOTerms) | GOTerms == "", ParentTerm, GOTerms)) %>%
  dplyr::rename(ChildTerm = GOTerms) -> GO_slim_edited

# Merge GO_results with GO_slim to add GOTerms column
GO_slim_edited %>%
  full_join(GO_results, by = c("module", "ChildTerm" = "GO.ID", "Ontology")) %>% 
  dplyr::rename(GO_slim_term = ParentTerm, GO_slim_term_description = Term.x, GO_child_term_description = Term.y) -> merged_results

write_csv(merged_results, "4_WGCNA/GO_enrichment_all_modules_allterms_withGOslim.csv")
```

```{r}
GO_results %>% 
  filter(module == "darkmagenta" | module == "mediumpurple3") %>%
  group_by(Ontology, module) %>%
  summarise(total_genes = sum(Significant))
```



# bubble plots for significant modules
```{r}
merged_results %>% 
  filter(module == "darkmagenta" | module == "mediumpurple3") -> sig_wgcna_modules

sig_wgcna_modules %>% 
  drop_na() %>% 
  dplyr::select(GO_slim_term:Ontology) %>% 
  distinct() %>% 
  write_csv("4_WGCNA/GO_slim_terms_sig_modules.csv")
```


```{r}
sig_wgcna_modules %>% 
  drop_na() %>% 
  dplyr::select(GO_slim_term:Ontology) %>% 
  distinct() %>% 
ggplot(., aes(x = fct_reorder(GO_slim_term_description, Count), y = Count, fill = Ontology)) +
  geom_bar(stat = "identity") +
  scale_fill_manual(values = c("BP" = "#8DD3C7",
                                "MF" = "#FFFFB3",
                                "CC" = "#BEBADA")) +
  labs(x = "Parent GO Term",
       y = "Count",
       fill = "Ontology") +
  theme_bw() +
  theme(
    axis.text.y = element_text(size = 9),
    legend.position = "bottom"
  ) +
  facet_wrap(~module, scales = "free") +
  coord_flip()
#ggsave("4_WGCNA/GO_slim_barplot_sig_modules.pdf", width = 10, height = 6)
```


```{r}
# Prepare data: select top terms per ontology
top_terms <- GO_results %>%
  filter(module == "darkmagenta" | module == "mediumpurple3") %>%
  group_by(Ontology, module) %>%
  arrange(weightFisher) %>%
  slice_head(n = 10) %>%  # Top 10 per ontology
  ungroup() %>%
  mutate(
    GeneRatio = Significant / Annotated,  # Calculate gene ratio
    NegLog10P = -log10(weightFisher),
    Term = fct_reorder(Term, NegLog10P)  # Order by significance
  )

# Create dot plot
top_terms %>% 
  filter(Ontology!="CC") %>% 
ggplot(., aes(x = GeneRatio, y = fct_reorder(Term, GeneRatio))) +
  geom_point(aes(size = Significant, color = NegLog10P)) +
  scale_color_gradient(low = "blue", high = "red", 
                       name = "-log10(P-value)") +
  scale_size_continuous(name = "Gene Count", range = c(1, 8)) +
  facet_grid(Ontology ~ module, scales = "free_y", space = "free_y") +
  labs(x = "Gene Ratio",
       y = "") +
  theme_bw() +
  theme(
    strip.text = element_text(size = 14, face = "bold"),
    axis.text.y = element_text(size = 14),
    panel.grid.major.y = element_line(color = "grey90"),
    panel.grid.minor = element_blank(),
    legend.position = "top"
  ) 
ggsave("4_WGCNA/GO_dotplot_sig_modules.pdf", width = 13, height = 8)
```
